requirements.txt
torchvision
torch
numpy
matplotlib
Pillow